Changing system actors. Improved scheduler. Changing the methods of sending messages to actors.
Adds a new type of actor - system actor (previously, they were not explicitly present).

Duplicates each method of sending a message to another actor. Adds version with a subscription to the state of the message (and to reply to messages) and without a subscription.

Improves scheduler. In addition to repeatedly actions, there are one shot actions.
CancellationToken replace to RepeatedlyActionToken for repeatedly actions and OneShotActionToken for one shot actions.
For OneShotActionToken and RepeatedlyActionToken add the ability to receive links to them and transfer them to other actors (isolates).

Changes to the data store in the actor, which were passed to him when he was created.

Updates documentation.
Updates and adds examples.
Adds tests.